+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Src filter for python now correctly excludes additional `srcExclude`s.
+
 ## [3.0.1] - 2021-11-25
 
 ### Fixed